setup: Set gid to match local machine's
DataLad's tests assume that the temporary directory is mounted in the
container and that it has the appropriate permissions on the files.
We already set the dl user's uid to match the local machine to avoid
permission issues.  Make the gid match as well.

This issue was revealed by test_create_sibling.test_target_ssh_inherit
failing in an attempt to use the Docker target in datalad-extensions:
https://github.com/datalad/datalad-extensions/runs/920594967